Rename the module `token_service`
This PR renames the module `token_service` to `token_provider` since the file contains the definition of the `TokenProvider` trait.
