+++ src/services.rs
@@ -4,5 +4,5 @@ pub mod base;
 pub mod observability;
 pub mod service_provider;
 pub mod token_decryption_service;
-pub mod token_service;
+pub mod token_provider;
 pub mod validation_service;
