+++ src/services/token_provider/external_identity.rs
